2 commit

+++ Client/Models/Car.cs
@@ -11,7 +11,7 @@ namespace Client.Models
         public int Id { get; set; }
         public string? Make { get; set; } 
         public string? Model { get; set; }
-        public ushort Year { get; set; } = 0;
+        public int Year { get; set; } = 0;
         public string? VIN { get; set; } 
         public string? Color { get; set; } 
     }
